changed list to int

+++ 0x14-more_singly_linked_lists/0-print_listint.c
@@ -11,7 +11,7 @@
 size_t print_listint(const listint_t *h)
 
 {
-	size_t insomnia = 0;
+	int insomnia = 0;
 
 	while (h != NULL)
 	{
